backend/requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
collection==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
drf-yasg==1.21.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
mysql-client==0.0.1
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.5.2
